15 compose api (#21)
* Moves index to .src folder

* moves resolvers

* Decomposes resolvers
