+++ package.json
@@ -7,7 +7,7 @@
     "import": "node -r dotenv/config ./import-data",
     "postinstall": "node ./import-data",
     "start": "node index.js",
-    "dev": "nodemon -r dotenv/config . -e graphql,js,json"
+    "dev": "nodemon -r dotenv/config ./src -e graphql,js"
   },
   "keywords": [],
   "author": "",
