+++ src/resolvers/index.js
@@ -1,0 +1,15 @@
+const Query = require("./Query");
+const Mutation = require("./Mutation");
+const Checkout = require("./Checkout");
+const Customer = require("./Customer");
+const Date = require("./Date");
+const Pet = require("./Pet");
+
+module.exports = {
+  Query,
+  Mutation,
+  Date,
+  Checkout,
+  Customer,
+  Pet
+};
